pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiogram>=3.26.0",
    "fastapi==0.116.1",
    "pydantic[email]==2.11.7",
    "pydantic-settings==2.10.1",
    "python-dotenv>=1.2.2",
    "sqlalchemy==2.0.43",
    "uvicorn[standard]==0.35.0",
    "python-jose[cryptography]==3.5.0",
    "passlib[bcrypt]==1.7.4",
    "email-validator==2.3.0",
    "bcrypt==3.2.0",
    "jinja2>=3.1.6",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "httpx>=0.28.1",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "ruff>=0.15.5",
]
